move map component to components directory
