+++ static/src/components/map/index.js
